Changed return value type in two functions to get rid of some warnings.
git-svn-id: https://subversion.gsi.de/fairroot/cbmroot/trunk@3298 0381ead4-6506-0410-b988-94b70fbc4730

+++ htrack/MiscLIB/src/conversionRoutines.cxx
@@ -1278,7 +1278,7 @@ unsigned int stoui(char* value, int radix) {
 
 	}
 	else
-		returnValue = (unsigned long) -1;
+		returnValue = (unsigned int) -1;
 
 	return returnValue;
 
@@ -1319,7 +1319,7 @@ unsigned int stoui(std::string value, int radix) {
 
 	}
 	else
-		returnValue = (unsigned long) -1;
+		returnValue = (unsigned int) -1;
 
 	return returnValue;
